[Calib] Move private header in to src directory
